+++ src/main/kotlin/simulation/BoxState.kt
@@ -4,8 +4,8 @@ data class BoxState(
     val position: Double,
     val velocity: Double,
     val mass: Long,
-    val width: Int,
-    val height: Int,
+    val width: Double,
+    val height: Double,
 ) {
     fun move(dt: Double): BoxState {
         return copy(position = position + velocity * dt)
